+++ common-library/src/main/java/com/jack/common/dto/response/WalletResponseDto.java
@@ -4,6 +4,7 @@ import lombok.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -12,6 +13,6 @@ public class WalletResponseDto implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
     private Long userId;
-    private double usdBalance;
-    private double btcBalance;
+    private BigDecimal usdBalance;
+    private BigDecimal btcBalance;
 }
