+++ common-library/src/main/java/com/jack/common/dto/response/WalletUpdateMessageDto.java
@@ -5,12 +5,13 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
 @Builder
 public class WalletUpdateMessageDto implements Serializable {
     private Long userId;
-    private double usdAmount;
-    private double btcAmount;
+    private BigDecimal usdAmount;
+    private BigDecimal btcAmount;
 }
